[REFACTOR]: 사용자 CRUD 및 보호자 CRUD 클래스 정리. 보호자 관련 메서드를 caregiver_crud.py로 이동하고, user_crud.py에서 관련 코드 제거.
