+++ users/user_crud.py
@@ -4,15 +4,13 @@
 사용자 정보, 프로필, 알림 설정에 대한 CRUD 작업
 """
 
-from typing import Optional, List
+from typing import Optional
 from sqlmodel import Session, select
 from sqlalchemy.orm import joinedload
 
 from common.postgres.crud.base import CRUDBase
 from .user import User, UserProfile, UserAlert
-from .caregivers.caregiver import Caregiver
-from .safety_area import SafetyArea
-from common.postgres.crud.base import OAuthPlatform
+from common.postgres.models import OAuthPlatform
 
 
 class CRUDUser(CRUDBase[User, None, None]):
@@ -90,42 +88,6 @@ class CRUDUserAlert(CRUDBase[UserAlert, None, None]):
             db.commit()
             db.refresh(alert)
         return alert
-
-
-class CRUDCaregiver(CRUDBase[Caregiver, None, None]):
-    """보호자 CRUD 작업"""
-    
-    def get_by_user_id(self, db: Session, user_id: int) -> List[Caregiver]:
-        """사용자 ID로 보호자 목록 조회"""
-        statement = select(Caregiver).where(Caregiver.user_id == user_id)
-        return db.exec(statement).all()
-    
-    def get_by_target_id(self, db: Session, target_id: int) -> List[Caregiver]:
-        """대상 ID로 보호자 목록 조회"""
-        statement = select(Caregiver).where(Caregiver.target_id == target_id)
-        return db.exec(statement).all()
-    
-    def create_caregiver_relationship(self, db: Session, user_id: int, target_id: int) -> Caregiver:
-        """보호자 관계 생성"""
-        caregiver = Caregiver(user_id=user_id, target_id=target_id)
-        db.add(caregiver)
-        db.commit()
-        db.refresh(caregiver)
-        return caregiver
-    
-    def delete_caregiver_relationship(self, db: Session, user_id: int, target_id: int) -> bool:
-        """보호자 관계 삭제"""
-        statement = select(Caregiver).where(
-            Caregiver.user_id == user_id,
-            Caregiver.target_id == target_id
-        )
-        caregiver = db.exec(statement).first()
-        if caregiver:
-            db.delete(caregiver)
-            db.commit()
-            return True
-        return False
-
 
 # CRUD 인스턴스 생성
 user_crud = CRUDUser(User)
